refactor: clean up so DPM usage is clearer
