+++ client/src/components/DataFiles/DataFilesForDPM/DataFilesForDPMBreadcrumbs.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+import { CMSBreadcrumbs } from '_common';
+
+/** To load CMS breadcrumbs with Data Files data */
+export default function DataFilesForDPMBreadcrumbs() {
+  return <CMSBreadcrumbs />;
+}
