+++ client/src/components/_common/CMSBreadcrumbs/index.js
@@ -1,0 +1,1 @@
+export { default } from './CMSBreadcrumbs';
